CI: Do not run the `step-tests` workflow on macOS for now.
Suposedly, 49460022b41cae29556d12870c447d1c540812fe had fixed the `step-tests` workflow on macOS, but it is still failing on the `macos-latest` runner for unknown reasons. It is not my fault, I have verified that the script works correctly on Ubuntu. However, it fails from time to time on GitHub, specially when the commit that triggers the workflow contains no changes to the workflow file.